Separate code generation module
