+++ src/graph_transpiler/webdnn/backend/code_generator/injectors/kernel_name_injector.py
@@ -1,7 +1,7 @@
 import hashlib
 from typing import Union
 
-from webdnn.backend.webgpu.injector import Tag, Injector
+from webdnn.backend.code_generator.injector import Tag, Injector
 from webdnn.graph.operator import Operator
