+++ src/graph_transpiler/webdnn/backend/webgpu/kernels/concat.py
@@ -2,9 +2,9 @@ from typing import List
 
 import numpy as np
 
+from webdnn.backend.code_generator.injectors.kernel_name_injector import KernelNameInjector
+from webdnn.backend.code_generator.injectors.meta_injector import MetaInjector
 from webdnn.backend.webgpu.allocator import MemoryLayout
-from webdnn.backend.webgpu.injectors.kernel_name_injector import KernelNameInjector
-from webdnn.backend.webgpu.injectors.meta_injector import MetaInjector
 from webdnn.backend.webgpu.kernel import GPUSize, Kernel
 from webdnn.graph.operators.concat import Concat
